Add files via upload
This is a C++ simple program where two string variables are declared globally, one for the users input and the other for storing the reversal.  The program will take the users input in the form of a string then confirms what the user entered.  Once the user is finished entering the string, a "FOR Loop" is used the reverse the user's string.