Rename ReverseString.cpp to Palindrome.cpp
This C++ program checks a string that the user inputs whether the string reversed is a palindrome (a string that is spelled the same backwards as forwards).  The program starts by first creating two empty string variables, one the store the users input and the other to store the reversed string.  Start by asking the user to input a string or sentence to be reversed and checked the string the user input needs to be looped through to get its reversal and this can be achieved using a "for loop" where the condition for the "for loop" is the initializer < the strings length; ever time this is true, the initializer increases by one and the index of the character is added to the reversed string variable, using an "if" statement to compare the two strings.

For the string to be reversed, the loop needs to increment up not down, for example:
STRING -> G + N + I + R+ T+ S